+++ Samples/ShowEvents/Program.cs
@@ -35,10 +35,13 @@ using IKaleidescape client = new KaleidescapeClient(new() {
 // hook-up event handlers
 client.HighlightedSelectionChanged += async delegate (object? sender, HighlightedSelectionChangedEventArgs args) {
     var details = await client.GetContentDetailsAsync(args.SelectionId);
-    Console.WriteLine($"=> Details: {details.Title} ({details.Year}) [{args.SelectionId}]");
+    Console.WriteLine($"=> Movie Details: {details.Title} ({details.Year}) [{args.SelectionId}]");
 };
 client.UiStateChanged += async delegate (object? sender, UiStateChangedEventArgs args) {
-    Console.WriteLine($"=> UI: screen={args.Screen}, dialog={args.Dialog}, popup={args.Popup}, saver={args.Saver}");
+    Console.WriteLine($"=> UI State: screen={args.Screen}, dialog={args.Dialog}, popup={args.Popup}, saver={args.Saver}");
+};
+client.MovieLocationChanged += async delegate(object? sender, MovieLocationEventArgs args) {
+    Console.WriteLine($"=> Movie Location: location={args.Location} [{args.LocationDescription}]");
 };
 
 // connect to device